backend/requirements.txt
Django==4.2.4
django-colorfield==0.9.0
djangorestframework==3.14.0
djoser==2.2.0
flake8==6.1.0
Pillow==10.0.0
psycopg==3.1.10
python-dotenv==1.0.0